Squashed commit for ease of rebasing
Break out of write dep checking loop on flush

Begin working on acceleration data structure

flailing

Add proptests for active job tracker

[muffled screaming]

test_dep* now passing

Fix more tests, not quite done yet

More unit test fixes

All tests now pass (I think)

Refactor job checking

Clean up proptest oracle

manually wrapping

Consolidate intervals whenever possible

Remove print statements

Remove reallocation

More proptests!

Don't track Empty jobs

More self-checking

More cleanup, add flush proptest

Comment a tricky part

Rename block map

Include live repairs in block-to-job dependency tracking

Fighting unit tests

Fix last test; update comments

Optimize split calculation

Code cleanup

Optimize iter_jobs

Minor cleanup

Simplify split insertion

Always reserve dependencies before inserting the job

Remove impacted_blocks from DownstairsIO

just a quick change

Fix downstairs

Less logging

More less logging

Remove Dependencies type to minimize diff

Further minimize diff

Fix edited comment

Simplify repair deps to a straight chain

comment updates

more comment

Fix copypasta

Explain BLOCKS_PER_EXTENT

More comment

Better naming for repair ID reservations

Explain ExtentLiveNoOp

Remove test of unsupported behavior

Store jobs unconditionally, to make assertion checking stronger

Remove outdated comments

Clarify comment

+++ upstairs/src/impacted_blocks.rs
@@ -296,15 +296,15 @@ pub fn extent_from_offset(
 
 pub fn extent_to_impacted_blocks(
     ddef: &RegionDefinition,
-    eid: u32,
+    eid: u64,
 ) -> ImpactedBlocks {
-    assert!(eid < ddef.extent_count());
+    assert!(eid < ddef.extent_count() as u64);
     let one = ImpactedAddr {
-        extent_id: eid as u64,
+        extent_id: eid,
         block: 0,
     };
     let two = ImpactedAddr {
-        extent_id: eid as u64,
+        extent_id: eid,
         block: ddef.extent_size().value - 1,
     };
     ImpactedBlocks::InclusiveRange(one, two)
